requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
numpy>=2.0,<3.0
sentence-transformers>=3.0,<4.0
daphne>=4.0,<5.0
whitenoise>=6.6,<7.0
